Move to lsp-zero for lsp configuration
